feat: refactor model input v2
