+++ jams-core/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod manager;
 pub mod model;
 pub mod model_store;
+mod pool;
+
 pub const FEATURE_VALUES_CAPACITY: usize = 125_000;
 pub const FEATURE_NAMES_CAPACITY: usize = 50;
